Refactor change
rename files

refactor readme.md

replace version name with sha for github actions

add sha for missing actions

chores

remove isInitialized property

delete unused data models

Remove jvm

Remove macos main

Remove wasm

Remove unused libs

remove wasmjs target in sample

replace manual set call with native setPersonProperties method

Add js target for sample

fixes

remove survey in config for this version

add a note to getAllFeatureFlags() method

remove super properties

remove unused code and ensure some atomic behaviours

address comments

add samples

fixes

fixes

fixes

fix capture exception

fixes
