+++ gradle/libs.versions.toml
@@ -3,13 +3,12 @@ kotlin = "2.3.0"
 agp = "8.13.0"
 compose-multiplatform = "1.10.0"
 androidx-activity-compose = "1.12.3"
-posthog-android = "3.30.0"
-posthog-ios = "3.38.0"
-posthog-js = "1.328.0"
+posthog-android = "3.51.0"
+posthog-ios = "3.62.0"
+posthog-js = "1.393.0"
 coroutines = "1.10.2"
 kotlinx-datetime = "0.7.1"
 kotlinx-serialization = "1.10.0"
-ktor = "3.4.0"
 mavenPublish = "0.35.0"
 spmforkmp = "1.4.7"
 
@@ -35,13 +34,6 @@ kotlinx-datetime = { group = "org.jetbrains.kotlinx", name = "kotlinx-datetime",
 # PostHog SDKs
 posthog-android = { group = "com.posthog", name = "posthog-android", version.ref = "posthog-android" }
 
-# Ktor (for pure Kotlin HTTP implementation option)
-ktor-client-core = { group = "io.ktor", name = "ktor-client-core", version.ref = "ktor" }
-ktor-client-content-negotiation = { group = "io.ktor", name = "ktor-client-content-negotiation", version.ref = "ktor" }
-ktor-serialization-kotlinx-json = { group = "io.ktor", name = "ktor-serialization-kotlinx-json", version.ref = "ktor" }
-ktor-client-okhttp = { group = "io.ktor", name = "ktor-client-okhttp", version.ref = "ktor" }
-ktor-client-darwin = { group = "io.ktor", name = "ktor-client-darwin", version.ref = "ktor" }
-ktor-client-js = { group = "io.ktor", name = "ktor-client-js", version.ref = "ktor" }
 
 # Testing
 kotlin-test = { group = "org.jetbrains.kotlin", name = "kotlin-test", version.ref = "kotlin-test" }
