+++ posthog-kmp/src/commonMain/kotlin/com/posthog/kmp/PostHogConfig.kt
@@ -1,0 +1,89 @@
+package com.posthog.kmp
+
+/**
+ * Configuration options for PostHog SDK initialization.
+ *
+ * @property apiKey Your PostHog project API key (required)
+ * @property host PostHog instance URL (defaults to US cloud)
+ * @property debug Enable debug logging
+ * @property captureApplicationLifecycleEvents Automatically capture app lifecycle events
+ * @property captureScreenViews Automatically capture screen/page views
+ * @property captureDeepLinks Capture deep link opens (Android only)
+ * @property sendFeatureFlagEvent Send events when feature flags are evaluated
+ * @property preloadFeatureFlags Preload feature flags on initialization
+ * @property flushAt Number of events to batch before sending
+ * @property flushIntervalSeconds Interval in seconds between automatic flushes
+ * @property maxQueueSize Maximum number of events to queue
+ * @property maxBatchSize Maximum events per batch
+ * @property optOut Start with analytics opted out
+ * @property personProfiles Person profile mode for feature flag targeting
+ * @property sessionRecording Enable session recording (platform dependent)
+ * @property autocapture Enable automatic event capture (platform dependent)
+ * @property featureFlagRequestTimeoutMs Timeout in milliseconds for feature flag requests
+ */
+public data class PostHogConfig(
+    val apiKey: String,
+    val host: String = "https://us.i.posthog.com",
+    val debug: Boolean = false,
+    val captureApplicationLifecycleEvents: Boolean = true,
+    val captureScreenViews: Boolean = false,
+    val captureDeepLinks: Boolean = true,
+    val sendFeatureFlagEvent: Boolean = true,
+    val preloadFeatureFlags: Boolean = true,
+    val flushAt: Int = 20,
+    val flushIntervalSeconds: Int = 30,
+    val maxQueueSize: Int = 1000,
+    val maxBatchSize: Int = 50,
+    val optOut: Boolean = false,
+    val personProfiles: PersonProfiles = PersonProfiles.IDENTIFIED_ONLY,
+    val sessionRecording: SessionRecordingConfig? = null,
+    val autocapture: Boolean = false,
+    val featureFlagRequestTimeoutMs: Int = 10000
+) {
+    public companion object {
+        /** PostHog US Cloud instance */
+        public const val HOST_US: String = "https://us.i.posthog.com"
+
+        /** PostHog EU Cloud instance */
+        public const val HOST_EU: String = "https://eu.i.posthog.com"
+    }
+}
+
+/**
+ * Person profile modes for feature flag targeting and person data.
+ */
+public enum class PersonProfiles {
+    /** Create person profiles for all users */
+    ALWAYS,
+
+    /** Only create profiles for identified users */
+    IDENTIFIED_ONLY,
+
+    /** Never create person profiles */
+    NEVER
+}
+
+/**
+ * Session recording configuration options.
+ *
+ * @property enabled Enable session recording
+ * @property maskAllTextInputs Mask all text input values
+ * @property maskAllImages Mask all images
+ * @property captureNetworkTelemetry Include network requests in recording
+ * @property captureLogs Capture console logs
+ * @property screenshot Enable screenshot mode instead of wireframe (Android/iOS experimental).
+ *                       When enabled, the SDK takes actual screenshots instead of wireframe representations.
+ *                       WARNING: Screenshots may contain sensitive information - ensure proper masking.
+ * @property captureLogcat Capture Android logcat output (Android only)
+ * @property debouncerDelayMs Delay in milliseconds for debouncing touch events (Android only)
+ */
+public data class SessionRecordingConfig(
+    val enabled: Boolean = true,
+    val maskAllTextInputs: Boolean = true,
+    val maskAllImages: Boolean = false,
+    val captureNetworkTelemetry: Boolean = true,
+    val captureLogs: Boolean = true,
+    val screenshot: Boolean = false,
+    val captureLogcat: Boolean = false,
+    val debouncerDelayMs: Long = 500L
+)
