+++ posthog-kmp/src/jsMain/kotlin/com/posthog/kmp/PostHogContext.js.kt
@@ -1,0 +1,14 @@
+package com.posthog.kmp
+
+/**
+ * JavaScript PostHog context.
+ * No platform-specific context is required for browser/Node.js.
+ */
+public actual class PostHogContext internal constructor(
+    @Suppress("unused") private val unit: Unit = Unit
+)
+
+/**
+ * Creates a PostHogContext for JavaScript.
+ */
+public actual fun PostHogContext(): PostHogContext = PostHogContext(Unit)
