--- posthog-kmp/src/jvmMain/kotlin/io/github/samuolis/posthog/PostHogContext.jvm.kt
@@ -1,14 +1,0 @@
-package io.github.samuolis.posthog
-
-/**
- * JVM PostHog context.
- * No platform-specific context is required for JVM.
- */
-public actual class PostHogContext internal constructor(
-    @Suppress("unused") private val unit: Unit = Unit
-)
-
-/**
- * Creates a PostHogContext for JVM.
- */
-public actual fun PostHogContext(): PostHogContext = PostHogContext(Unit)
