+++ sample/src/androidMain/kotlin/com/posthog/kmp/sample/MainActivity.kt
@@ -1,0 +1,19 @@
+package com.posthog.kmp.sample
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import com.posthog.kmp.PostHogContext
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        enableEdgeToEdge()
+        setContent {
+            // Pass the Android Application context to PostHog
+            App(postHogContext = PostHogContext(application))
+        }
+    }
+}
