+++ sample/src/androidMain/kotlin/com/posthog/kmp/sample/Platform.android.kt
@@ -1,0 +1,3 @@
+package com.posthog.kmp.sample
+
+actual fun getPlatformName(): String = "Android"
