--- sample/src/androidMain/kotlin/io/github/samuolis/posthog/sample/MainActivity.kt
@@ -1,19 +1,0 @@
-package io.github.samuolis.posthog.sample
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import io.github.samuolis.posthog.PostHogContext
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        enableEdgeToEdge()
-        setContent {
-            // Pass the Android Application context to PostHog
-            App(postHogContext = PostHogContext(application))
-        }
-    }
-}
