--- sample/src/androidMain/kotlin/io/github/samuolis/posthog/sample/Platform.android.kt
@@ -1,3 +1,0 @@
-package io.github.samuolis.posthog.sample
-
-actual fun getPlatformName(): String = "Android"
