+++ sample/src/iosMain/kotlin/com/posthog/kmp/sample/MainViewController.kt
@@ -1,0 +1,8 @@
+package com.posthog.kmp.sample
+
+import androidx.compose.ui.window.ComposeUIViewController
+import com.posthog.kmp.PostHogContext
+
+fun MainViewController() = ComposeUIViewController {
+    App(postHogContext = PostHogContext())
+}
