+++ sample/src/iosMain/kotlin/com/posthog/kmp/sample/Platform.ios.kt
@@ -1,0 +1,5 @@
+package com.posthog.kmp.sample
+
+import platform.UIKit.UIDevice
+
+actual fun getPlatformName(): String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
