+++ sample/src/jsMain/kotlin/com/posthog/kmp/sample/Platform.js.kt
@@ -1,0 +1,3 @@
+package com.posthog.kmp.sample
+
+actual fun getPlatformName(): String = "Web Browser (JS)"
