--- sample/src/wasmJsMain/resources/index.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>PostHog KMP Sample</title>
-    <style>
-        html, body {
-            margin: 0;
-            padding: 0;
-            width: 100%;
-            height: 100%;
-            overflow: hidden;
-        }
-        #ComposeTarget {
-            width: 100%;
-            height: 100%;
-        }
-    </style>
-</head>
-<body>
-    <canvas id="ComposeTarget"></canvas>
-    <script src="sample.js"></script>
-</body>
-</html>
